+++ splice_machine/src/test/java/com/splicemachine/derby/impl/sql/execute/operations/DistinctScalarAggregateOperationIT.java
@@ -165,7 +165,7 @@ public class DistinctScalarAggregateOperationIT extends SpliceUnitTest {
 
     @Test
     public void testDistinctCountOrderBy() throws Exception {
-        Assert.assertEquals(4L, methodWatcher.query("select count(distinct score) from " + this.getPaddedTableReference("ORDERSUMMARY") + "order by 1"));
+        Assert.assertEquals(4L,(long) methodWatcher.query("select count(distinct score) from " + this.getPaddedTableReference("ORDERSUMMARY") + "order by 1"));
     }
 
     @Test
